Seperate sponsors list to new file
